# src/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "",
    "version": "v1",
    "title": "Flask API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
        {
        "url": "/"
        }
    ],
   "components": {
     "securitySchemes": {
       "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
      }
  },
...
        "schemas": {
            "Users": {
                "type": "object",
                "properties": {
                    "username": {
                    "type": "string"
                    },
                    "email": {
                    "type": "string"
                    }
                }
            },
...
            "Register": {
                "type": "object",
                "properties": {
                    "username": {
                    "type": "string"
                    },
                    "email": {
                    "type": "string"
                    },
                    "password": {
                    "type": "string"
                    }
                }
            },
...
          "Add": {
                "type": "object",
                "properties": {
                    "userId": {
                    "type": "integer"
                    },
                    "productId": {
                    "type": "integer"
                    },
                    "quantity": {
                    "type": "integer"
                    }
                }
            },
...
          "Carts": {
                "type": "object",
                "properties": {
                    "ProductID": {
                    "type": "integer"
                    },
                    "Quantity": {
                    "type": "integer"
                    }
                }
            },
          "Id":{
            "type" : "integer",
            "properties" : {
                "name" : {
                    "type" : "integer"
                }
            }
        },
...
           "Address": {
                "type": "object",
                "properties": {
                  "Address": {
                    "type": "string"
                    }
                }
            },
          "Send": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                    },
                  "address": {
                    "type": "string"
                    },
                  "latitude": {
                    "type": "string"
                  },
                  "longitude":{
                    "type": "string"
                  }
                }
            },
...
          "update":{
              "type": "object",
              "properties" :{
                "Id": {
                    "type": "integer"
                    },
                  "UserId": {
                    "type": "integer"
                  },
                  "ProductId": {
                    "type": "integer"
                  },
                  "Quantity": {
                    "type": "integer"
                  }
              }
            },
...
              "RemoveFav": {
                "type": "object",
              "properties" :{
                  "product_id": {
                    "type": "integer"
                  }
              }
              }
        }
    },
     "paths": {
...
       "/api/v1/favorites/GetFavorites": {
         "get": {
        "tags": [
            "Favorites"
        ],
        "summary": "Get Users favorites",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "This endpoint retrieves user information based on the user's token information.",
        "responses": {
            "200": {
                "description": "Success",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Users"
                        }
                    }
                }
            },
            "404": {
                "description": "NOT FOUND",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Users"
                        }
                    }
                }
            }
        }
    }
},
       "/api/v1/favorites/AddFavorite": {
         "post": {
            "tags": [
                "Favorites"
            ],
            "summary": "Register User",
            "security": [
          {
            "bearerAuth": []
          }
        ],
            "requestBody": {
            "description": "Post New todo To Database",
            "required": true,
            "content": {
                "application/json": {
                "schema": {
                    "$ref": "#/components/schemas/Register"
                    }
                }
            }
            },
            "responses": {
                "200": {
                    "description": "OK",
                    "schema": {
                    "$ref": "#/components/schemas/Users"
                    }
                },
                "404": {
                    "description": "NOT FOUND",
                    "schema": {
                    "$ref": "#/components/schemas/Users"
                    }
            }
            }
            }
       },
       "/api/v1/favorites/RemoveFavorite": {
         "delete": {
           "tags": [
             "Favorites"
           ],
           "security": [
          {
            "bearerAuth": []
          }
        ],
           "requestBody": [
             {
               "description": "product id to delete",
               "required": true,
               "content": {
                 "application/json": {
                   "schema": {
                     "$ref": "#/components/schemas/RemoveFav"
                   }
                 }
                 }
               }
           ],
           "summary": "Remove product from favorites",
           "responses": {
             "200": {
               "description": "OK",
               "schema": {
                 "$ref": "#/components/schemas/Users"
               }
             },
             "404": {
               "description": "NOT FOUND",
               "schema": {
                 "$ref": "#/components/schemas/Users"
               }
             }
           }
         }
       },
...
       "/api/v1/carts/GetAll": {
          "get" :{
                "tags": [
                    "Carts"
                ],
                "summary": "Get All ",
            "security": [
          {
            "bearerAuth": []
          }
        ],
                "responses": {
                "200": {
                    "description": "Success",
                    "schema": {
                    "$ref": "#/components/schemas/Carts"
                    }
                },
                "404": {
                    "description": "NOT FOUND",
                    "schema": {
                    "$ref": "#/components/schemas/Carts"
                    }
            }
            }
            }
       },
        "/api/v1/carts/add": {
            "post": {
              "tags": [
                "Carts"
              ],
              "summary": "Add",
              "security": [
          {
            "bearerAuth": []
          }
        ],
              "requestBody": {
              "description": "Post New todo To Database",
              "required": true,
              "content": {
                  "application/json": {
                  "schema": {
                      "$ref": "#/components/schemas/Add"
                     }
                 }
              }
            },
            "responses": {
                "200": {
                    "description": "OK",
                    "schema": {
                    "$ref": "#/components/schemas/Add"
                    }
                },
                "404": {
                    "description": "NOT FOUND",
                    "schema": {
                    "$ref": "#/components/schemas/Add"
                    }
            }
            }
            }
       },
       "/api/v1/carts/update": {
         "put": {
          "tags": [
            "Carts"
          ],
          "summary": "Update",
           "security": [
          {
            "bearerAuth": []
          }
        ],
          "requestBody": {
            "description": "Update Existing todo On Database",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/update"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/components/schemas/Users"
              }
            },
            "404": {
              "description": "NOT FOUND",
              "schema": {
                "$ref": "#/components/schemas/Users"
              }
            }
          }
        }
        },
       "/api/v1/carts/remove": {
         "delete": {
           "tags": [
             "Carts"
           ],
           "security": [
          {
            "bearerAuth": []
          }
        ],
           "parameters": [
             {
               "name": "CartId",
               "in": "path",
               "description": "cart id to delete",
               "required": true,
               "type": "integer",
               "schema": {
                 "$ref": "#/components/schemas/Id"
               }
             }
           ],
           "summary": "Delete todo",
           "responses": {
             "200": {
               "description": "OK",
               "schema": {
                 "$ref": "#/components/schemas/Users"
               }
             },
             "404": {
               "description": "NOT FOUND",
               "schema": {
                 "$ref": "#/components/schemas/Users"
               }
             }
           }
         }
       },
...
        "/api/v1/map/convert": {
          "get" :{
                "tags": [
                    "Map"
                ],
                "summary": "Get Address",
            "parameters": [
          {
            "name": "address",
            "in": "query",
            "description": "Address for convert to lat&lng ",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
                "responses": {
                "200": {
                    "description": "Success",
                    "schema": {
                    "$ref": "#/components/schemas/Address"
                    }
                },
                "404": {
                    "description": "NOT FOUND",
                    "schema": {
                    "$ref": "#/components/schemas/Address"
                    }
            }
            }
            }
       },
      "/api/v1/map/send": {
  "post": {
    "tags": [
      "Map"
    ],
    "summary": "Send request",
    "requestBody": {
      "description": "Data to be sent to the server",
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Send"
            }
          }
        }
      }
    },
    "responses": {
      "200": {
        "description": "OK",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Addresses sent to server successfully"
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "NOT FOUND",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Failed to send addresses to server"
                }
              }
            }
          }
        }
      }
    }
  }
},
 "/api/v1/map/get": {
  "get": {
    "tags": [
      "Map"
    ],
    "summary": "Fetch location data",
    "responses": {
      "200": {
        "description": "OK",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "latitude": {
                    "type": "number",
                    "example": 40.712776
                  },
                  "longitude": {
                    "type": "number",
                    "example": -74.005974
                  }
                }
              }
            },
            "examples": {
              "example-1": {
                "summary": "Example Response",
                "value": [
                  {
                    "latitude": 40.712776,
                    "longitude": -74.005974
                  },
                  {
                    "latitude": 34.052235,
                    "longitude": -118.243683
                  }
                ]
              }
            }
          }
        }
      },
      "500": {
        "description": "Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An error occurred while fetching locations"
                }
              }
            }
          }
        }
      }
    }
  }
},
...
        }
  }